Move remaining formatting code to Formatter
